refactor topiclayout & tooic banner
